// src/app/models/infobasica/tableDadosCTB.model.ts
export interface tableDadosCTB {
    bbP_DadosCTBID: string;
    tipoFJ: string;
    cnpj: string;
    cpf: string;
    codigoCNAE: string;
    inscricaoEstadual: string;
    inscricaoEstadual_Sub_Tributario: string;
    inscricaoMunicipal: string;
    indicador_natureza_pessoa_juridicial: string;
    indicador_Tipo_Atividade_Preponderante: string;
    ind_Apuracao_Contribuicoes_Creditos: string;
    indicador_Apuracao_IPI: string;
    indicador_Tipo_Sociedade_Cooperativa: string;
    tributacao_lucro: string;
    qualificacao_Empresa: string;
    tipo_declarante: string;
    valor_Acumulado_Imposto_Retido_Fonte_CR: string;
    valor_Acumulado_Imposto_Retido_Fonte_CP: string;
    selected?: boolean;
}  